+++ CHANGELOG.md
@@ -5,11 +5,32 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+
+## Version [3.4.0] - 2024-08-05 🚀
+
+
+### Fixed 🛠️
+
+- [Issue #27](https://github.com/I-am-PUID-0/DMB/issues/27) DATABASE_URL was not being set correctly for the Riven frontend 🐛
+- healthcheck: Fixed healthcheck to properly check for Riven frontend setup completion 🩺
+
+### Changed 🔄
+
+- Riven setup: Refactored backend and frontend setup to use ProcessHandler for consistent logging 🔄
+
+### Added ✨
+
+- [Issue #9](https://github.com/I-am-PUID-0/DMB/issues/9) Obfuscate sensitive data in logs
+- Riven backend: UPDATERS_PLEX_ADDRESS linked to PLEX_ADDRESS 🔄
+- Riven backend: UPDATERS_PLEX_TOKEN linked to PLEX_TOKEN 🔄
+
+
 ## Version [3.3.2] - 2024-08-03 🚀
 
 ### Fixed 🛠️
 
-[Issue #27](https://github.com/I-am-PUID-0/DMB/issues/27) Riven frontend needed a database connection to function properly 🐛
+- [Issue #27](https://github.com/I-am-PUID-0/DMB/issues/27) Riven frontend needed a database connection to function properly 🐛
 
 ### Added ✨
 
@@ -18,6 +39,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed 🔄
 
 - GITHUB_TOKEN: Can be added to the environment variables to allow for repository downloads without rate limits 🔄
+
 
 ## Version [3.3.1] - 2024-08-01 🚀